--- core/src/main/java/de/danoeh/antennapod/core/dialog/StatusListener.java
@@ -1,7 +1,0 @@
-package de.danoeh.antennapod.core.dialog;
-
-public interface StatusListener {
-    void onActionSuccess();
-
-    void onActionFailure(Throwable throwableError);
-}
